# docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuickWiki API
  version: 1.0.0
  description: API for QuickWiki
...
paths:
...
  /lecture:
    post:
      tags:
        - lecture
      summary: 新しい講習会情報の作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/lectureWithoutId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/lecture'
  /lecture/{lectureId}:
    get:
      tags:
        - lecture
      summary: 1つの講習会情報の取得
      parameters:
        - name: lectureId
          in: path
          required: true
          schema:
            type: string
          description: 講習会ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/lecture'
        '400':
          description: 講習会IDが整数でないと思われる
        '404':
          description: 講習会IDに対応する講習会が見つからない
...
components:
  schemas:
...
    lecture:
      type: object
      properties:
        id:
          type: integer
          example: 123
        title:
          type: string
          example: Introduction to Computer Science
        content:
          type: string
          example: 資料へのリンク(URL)のみ
        folderpath:
          type: string
          example: /School/ComputerScience
...
    lectureWithoutId:
      type: object
      properties:
        title:
          type: string
          example: Introduction to Computer Science
        content:
          type: string
          example: マークダウン形式 資料へのリンク
        folderpath:
          type: string
          example: /School/ComputerScience
...